Removed and integrated it into ViewWrapper
