+++ src/domain/collaboration/whiteboard/WhiteboardsManagement/WhiteboardsRtManagementViewWrapper.tsx
@@ -1,11 +1,22 @@
 import React, { FC } from 'react';
 import WhiteboardRtActionsContainer from '../containers/WhiteboardRtActionsContainer';
-import { AuthorizationPrivilege, WhiteboardRtDetailsFragment } from '../../../../core/apollo/generated/graphql-schema';
-import WhiteboardRtManagementView, {
-  ActiveWhiteboardIdHolder,
-  WhiteboardNavigationMethods,
-} from './WhiteboardRtManagementView';
+import { AuthorizationPrivilege, WhiteboardRtContentFragment, WhiteboardRtDetailsFragment } from '../../../../core/apollo/generated/graphql-schema';
 import { JourneyTypeName } from '../../../journey/JourneyTypeName';
+import WhiteboardRtContentContainer from '../containers/WhiteboardRtContentContainer';
+import WhiteboardRtDialog from '../WhiteboardDialog/WhiteboardRtDialog';
+import { useFullscreen } from '../../../../core/ui/fullscreen/useFullscreen';
+import FullscreenButton from '../../../../core/ui/button/FullscreenButton';
+import ShareButton from '../../../shared/components/ShareDialog/ShareButton';
+import { BlockTitle } from '../../../../core/ui/typography';
+import useWhiteboardRtContentUpdatePolicy from '../whiteboardRt/contentUpdatePolicy/WhiteboardRtContentUpdatePolicy';
+import WhiteboardShareSettings from '../share/WhiteboardShareSettings';
+
+export interface ActiveWhiteboardIdHolder {
+  whiteboardNameId?: string;
+}
+export interface WhiteboardNavigationMethods {
+  backToWhiteboards: () => void;
+}
 
 export interface WhiteboardsRtManagementViewWrapperProps extends ActiveWhiteboardIdHolder, WhiteboardNavigationMethods {
   journeyTypeName: JourneyTypeName;
@@ -29,9 +40,14 @@ const WhiteboardsRtManagementViewWrapper: FC<WhiteboardsRtManagementViewWrapperP
   readOnlyDisplayName,
   ...whiteboardsState
 }) => {
-  if (!calloutId) {
-    return null;
-  }
+  const { fullscreen, setFullscreen } = useFullscreen();
+
+  const handleCancel = () => {
+    backToWhiteboards();
+    if (fullscreen) {
+      setFullscreen(false);
+    }
+  };
 
   // Todo: need to decide who can edit what whiteboards, for now tie to UpdateContent. May need to extend the information on a Whiteboard
   // to include who created it etc.
@@ -40,29 +56,63 @@ const WhiteboardsRtManagementViewWrapper: FC<WhiteboardsRtManagementViewWrapperP
   const hasUpdatePrivileges = authorization?.myPrivileges?.includes(AuthorizationPrivilege.Update);
   const hasUpdateContentPrivileges = authorization?.myPrivileges?.includes(AuthorizationPrivilege.UpdateContent);
 
+  const contentUpdatePolicyProvided = useWhiteboardRtContentUpdatePolicy({
+    whiteboardId: whiteboard?.id,
+    skip: !hasUpdatePrivileges,
+  });
+
+  if (!calloutId) {
+    return null;
+  }
+
+
   return (
     <WhiteboardRtActionsContainer>
       {(_, actionsState, actions) => (
-        <WhiteboardRtManagementView
-          entities={{
-            whiteboard,
-            calloutId,
-            whiteboardNameId,
-            contextSource: journeyTypeName,
+        <WhiteboardRtContentContainer whiteboardId={whiteboard?.id}>
+          {entities => {
+            return (
+              <WhiteboardRtDialog
+                entities={{
+                  whiteboard: entities.whiteboard as WhiteboardRtContentFragment & WhiteboardRtDetailsFragment,
+                }}
+                actions={{
+                  onCancel: handleCancel,
+                  onUpdate: actions.onUpdate,
+                }}
+                options={{
+                  canEdit: hasUpdateContentPrivileges,
+                  show: Boolean(whiteboardNameId),
+                  fixedDialogTitle: hasUpdatePrivileges && !readOnlyDisplayName ? undefined : (
+                    <BlockTitle display="flex" alignItems="center">
+                      {whiteboard?.profile.displayName}
+                    </BlockTitle>
+                  ),
+                  fullscreen,
+                  headerActions: (
+                    <>
+                      <ShareButton url={whiteboardShareUrl} entityTypeName="whiteboard" disabled={!whiteboardShareUrl}>
+                        {hasUpdatePrivileges && (
+                          <WhiteboardShareSettings
+                            createdBy={entities.whiteboard?.createdBy}
+                            journeyTypeName={journeyTypeName}
+                            {...contentUpdatePolicyProvided}
+                          />
+                        )}
+                      </ShareButton>
+                      <FullscreenButton />
+                    </>
+                  ),
+                }}
+                state={{
+                  ...whiteboardsState,
+                  ...actionsState,
+                }}
+              />
+            );
           }}
-          actions={actions}
-          state={{
-            ...whiteboardsState,
-            ...actionsState,
-          }}
-          options={{
-            canUpdate: hasUpdatePrivileges,
-            canUpdateContent: hasUpdateContentPrivileges,
-            canUpdateDisplayName: hasUpdatePrivileges && !readOnlyDisplayName,
-            shareUrl: whiteboardShareUrl,
-          }}
-          backToWhiteboards={backToWhiteboards}
-        />
+        </WhiteboardRtContentContainer>
+
       )}
     </WhiteboardRtActionsContainer>
   );
